pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "obs-relay"
version = "1.0.0"
description = "Remote OBS control relay with REST API, WebSocket bridge, and TouchOSC support"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "obs-websocket-py>=1.0.0",
    "python-osc>=1.8.3",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "aiofiles>=23.2.1",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "python-multipart>=0.0.9",
    "PyYAML>=6.0",
    "apscheduler>=3.10.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
]
build = [
    "pyinstaller>=6.6.0",
]

[project.scripts]
obs-relay = "obs_relay.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["obs_relay*"]

[tool.ruff]
line-length = 100
target-version = "py310"
